Refactor ambiguous ID naming scheme

+++ TRT2API/Data/Models/MatchMap.cs
@@ -3,9 +3,9 @@
 public class MatchMap
 {
 	public int Id { get; set; }
-	public long MatchId { get; set; }
-	public long PlayerId { get; set; }
-	public long MapId { get; set; }
+	public int MatchId { get; set; }
+	public int PlayerId { get; set; }
+	public int MapId { get; set; }
 	public string Action { get; set; } = string.Empty;
 	public int OrderInMatch { get; set; }
 }
