+++ TRT2API/Data/Models/MatchPlayer.cs
@@ -3,7 +3,7 @@
 public class MatchPlayer
 {
 	public int Id { get; set; }
-	public long MatchId { get; set; }
-	public long PlayerId { get; set; }
+	public int MatchId { get; set; }
+	public int PlayerId { get; set; }
 	public int? Score { get; set; }
 }
